+++ src/elements/GeoTrimmedContent/GeoTrimmedContent.vue
@@ -1,9 +1,29 @@
+<template>
+  <span
+    v-on-resize="reloadRequiredWidth"
+    :class="`geo-trimmed-content${cssSuffix}`"
+    @mouseenter="isHovered = true"
+    @mouseleave="isHovered = false"
+  >
+    <geo-tooltip
+      :visible="isTooltipVisible"
+    >
+      <div ref="tooltipContent" />
+    </geo-tooltip>
+    <span
+      ref="content"
+      class="geo-trimmed-content__content"
+    >
+      <!-- @slot Use this slot to customize content displayed -->
+      <slot />
+    </span>
+  </span>
+</template>
+
 <script>
 import OnResize from '../../directives/GeoOnResize'
 import Tooltip from '../../directives/Tooltip'
 import cssSuffix from '../../mixins/cssModifierMixin'
-import counterFactory from '../../utils/counterFactory'
-const getNextInstanceId = counterFactory()
 
 export default {
   name: 'GeoTrimmedContent',
@@ -15,8 +35,7 @@ export default {
     return {
       isHovered: false,
       containerWidth: null,
-      contentWidth: null,
-      tooltipHTML: null
+      contentWidth: null
     }
   },
   computed: {
@@ -28,27 +47,20 @@ export default {
       const contentWidth = this.contentWidth || 0
       const containerWidth = this.containerWidth || 0
       return contentWidth > containerWidth
-    },
-
-    idTooltipContentNode () {
-      return `geo-trimmed-content--${getNextInstanceId()}`
     }
   },
-  beforeMount () {
-    this.addTooltipHTMLContentNode()
+  watch: {
+    isTooltipVisible () {
+      this.reloadTooltipContent()
+    }
   },
   mounted () {
-    this.addTooltipHTMLContentNode()
     this.reloadRequiredWidth()
     this.reloadTooltipContent()
   },
   updated () {
-    this.addTooltipHTMLContentNode()
     this.reloadRequiredWidth()
     this.reloadTooltipContent()
-  },
-  beforeDestroy () {
-    this.removeTooltipHTMLContentNode()
   },
   methods: {
     reloadRequiredWidth () {
@@ -70,61 +82,10 @@ export default {
     reloadTooltipContent () {
       if (!this.isContentTrimmed) return
       if (!this.$refs.content) return
+      if (!this.$refs.tooltipContent) return
 
-      this.tooltipHTML = this.$refs.content.innerHTML
-      const existingElement = document.getElementById(this.idTooltipContentNode)
-      existingElement.innerHTML = this.tooltipHTML
-    },
-
-    addTooltipHTMLContentNode () {
-      const element = document.getElementById(this.idTooltipContentNode)
-      if (element) return
-
-      const newElement = document.createElement('div')
-      newElement.setAttribute('id', this.idTooltipContentNode)
-      document.body.appendChild(newElement)
-    },
-
-    removeTooltipHTMLContentNode () {
-      if (!this.isContentTrimmed) return
-
-      const element = document.getElementById(this.idTooltipContentNode)
-      if (element) element.parentNode.removeChild(element)
+      this.$refs.tooltipContent.innerHTML = this.$refs.content.innerHTML
     }
-  },
-  render (createElement) {
-    const directives = [{
-      name: 'on-resize',
-      value: this.reloadRequiredWidth
-    }]
-
-    if (this.isContentTrimmed) {
-      directives.push({
-        name: 'tooltip',
-        value: {
-          visible: this.isTooltipVisible,
-          html: this.idTooltipContentNode
-        },
-        modifiers: {
-          top: true,
-          notrigger: true
-        }
-      })
-    }
-
-    return createElement('span', {
-      directives,
-      class: `geo-trimmed-content${this.cssSuffix}`,
-      on: {
-        mouseenter: () => { this.isHovered = true },
-        mouseleave: () => { this.isHovered = false }
-      }
-    }, [
-      createElement('span', {
-        class: 'geo-trimmed-content__content',
-        ref: 'content'
-      }, this.$slots.default)
-    ])
   }
 }
 </script>